new file:   README.md 	modified:   backend/logs/app.log 	modified:   backend/src/models/models.py 	modified:   backend/src/services/Chatbot.py 	modified:   backend/src/services/__pycache__/Chatbot.cpython-310.pyc 	modified:   backend/src/utils/__pycache__/file_operations.cpython-310.pyc 	modified:   frontend/gui/__pycache__/streamlit.cpython-310.pyc 	modified:   frontend/gui/streamlit.py

+++ backend/src/models/models.py
@@ -19,7 +19,7 @@ class Debt(Base):
     __tablename__ = 'debts'
 
     id = Column(Integer, primary_key=True)
-    total_debt = Column(Integer)
+    total_debt = Column(Float)
     maximum_period_months = Column(Integer)
     minimum_accepted_payment = Column(Float)
     user_email = Column(String, ForeignKey('users.email'))
